Update the image field value after an upload.o
